ENH: octave: Add the parstyle option.
The parstyle options controls how parameters are passed to the vector
field and Jacobian functions.  It accepts three values: 'global', 'list'
and 'vector'.

The default, 'global', maintains the existing behavior.  The parameters
are stored in a global variable.

The options 'list' and 'vector' have the same meaning as for the matlab
command.  'list' passes all the parameters as separate arguments.
'vector' passes a single vector that holds all the parameters.

When parstyle is 'list' or 'vector', the demo code uses anonymous
functions (i.e. closures) in the call to lsode.