ENH: Add version option for the dde_solver command.
In 2017, I created a modified version of dde_solver_m.f90 that
fixed some issues with function interfaces.  The new code now uses
assumed-shape arrays consistently.  This commit adds an option to
the dde_solver command that allows the user to select which version
of dde_solver to generate code for, the old version (version 1) or
the new version (version 2).